AoC 2023/14

+++ rust/src/advent_of_code_2023/advent_2023_3.rs
@@ -3,7 +3,7 @@ use itertools::Itertools;
 use regex::Regex;
 use std::fmt::{Debug, Display};
 
-fn read_maps() -> (Vec<Vec<char>>, Vec<(i32, i32, i32, i32)>) {
+fn read_maps() -> (Vec<Vec<char>>, Vec<(i64, i64, i64, i64)>) {
     let num_pattern = Regex::new(r"(\d+)").unwrap();
     let lines = read_lines("advent_2023/3.txt");
     let c = lines.iter().map(|l| l.chars().collect_vec()).collect_vec();
@@ -13,10 +13,10 @@ fn read_maps() -> (Vec<Vec<char>>, Vec<(i32, i32, i32, i32)>) {
         .flat_map(|(y, l)| {
             num_pattern.find_iter(l).map(move |m| {
                 (
-                    m.start() as i32,
-                    y as i32,
-                    m.as_str().len() as i32,
-                    m.as_str().parse::<i32>().unwrap(),
+                    m.start() as i64,
+                    y as i64,
+                    m.as_str().len() as i64,
+                    m.as_str().parse::<i64>().unwrap(),
                 )
             })
         })
@@ -24,12 +24,12 @@ fn read_maps() -> (Vec<Vec<char>>, Vec<(i32, i32, i32, i32)>) {
     (c, nums)
 }
 
-pub fn solve_1() -> i32 {
+pub fn solve_1() -> i64 {
     let (c, nums) = read_maps();
     nums.iter()
         .filter(|(x, y, l, v)| {
             (-1..=*l).cartesian_product(-1..=1).any(|(dx, dy)| {
-                if let Some((mx, my)) = map_add(*x, *y, dx, dy, c[0].len() as i32, c.len() as i32) {
+                if let Some((mx, my)) = map_add((*x, *y), (dx, dy), c[0].len() as i64, c.len() as i64) {
                     let mv = c[my as usize][mx as usize];
                     return !mv.is_numeric() && mv != '.';
                 }
@@ -37,15 +37,15 @@ pub fn solve_1() -> i32 {
             })
         })
         .map(|(_, _, _, v)| *v)
-        .sum::<i32>()
+        .sum::<i64>()
 }
 
-pub fn solve_2() -> i32 {
+pub fn solve_2() -> i64 {
     let (c, nums) = read_maps();
     let mut gears = vec![vec![(0, 1); c[0].len()]; c.len()];
     for (x, y, l, v) in &nums {
         for (dx, dy) in (-1..=*l).cartesian_product(-1..=1) {
-            if let Some((mx, my)) = map_add(*x, *y, dx, dy, c[0].len() as i32, c.len() as i32) {
+            if let Some((mx, my)) = map_add((*x, *y), (dx, dy), c[0].len() as i64, c.len() as i64) {
                 if c[my as usize][mx as usize] == '*' {
                     let g = &mut gears[my as usize][mx as usize];
                     *g = (g.0 + 1, g.1 * v)
@@ -53,5 +53,5 @@ pub fn solve_2() -> i32 {
             }
         }
     }
-    gears.iter().flatten().filter_map(|e| (e.0 == 2).then(|| e.1)).sum::<i32>()
+    gears.iter().flatten().filter_map(|e| (e.0 == 2).then(|| e.1)).sum::<i64>()
 }
